requirements.txt
dlib==19.24.4
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.4.0
setuptools==75.1.0